requirements.txt
requests
pillow
google-auth-oauthlib
google-auth-httplib2
google-api-python-client